Remove deprecated items.
