+++ protocols/gossipsub/Cargo.toml
@@ -31,7 +31,6 @@ quick-protobuf-codec = { workspace = true }
 hex_fmt = "0.3.0"
 regex = "1.8.1"
 serde = { version = "1", optional = true, features = ["derive"] }
-thiserror = "1.0"
 wasm-timer = "0.2.5"
 instant = "0.1.11"
 void = "1.0.2"
