+++ protocols/gossipsub/src/behaviour/tests.rs
@@ -21,7 +21,7 @@
 // Collection of tests for the gossipsub network behaviour
 
 use super::*;
-use crate::subscription_filter_priv::WhitelistSubscriptionFilter;
+use crate::subscription_filter::WhitelistSubscriptionFilter;
 use crate::transform::{DataTransform, IdentityTransform};
 use crate::types::FastMessageId;
 use crate::ValidationError;
